aws_resources/benchmarking/requirements.txt
ultralytics
super-gradients
pillow
torch
torchvision
